Change routing to use app-state
This profoundly changes the way routes are handled. Previously
routes update was handled separately from app-state update by
calling setState on the root component.

Now instead the router updates the app-state with the context of
the resolved route and the root component gets re-rendered automatically.

This has a number of benefits, namely it removes the need for route
initialisers, which can now simply be state observers (and as such, they
can be asynchronous and still work isomorphically). It also decouples
routing from the application module even more.